Refactor geoMapMenuItem
