+++ client/pages/Geo/GeoMap/GeoMapMenuItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './geoMapMenuItem'
