--- client/pages/Geo/GeoMap/GeoMapMenuItemHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './geoMapMenuItemHeader'
